Add files via upload
This is the project which is developed with the help of latest technologies such as HTML, CSS, and JavaScript. The main moto of this project is to demonstrate my Front End skills. In this I have used some of the music bytes which are used to make sounds based on the type we have selected. I have also fixed the tempo which is used to adjust according to our requirement.